+++ buildpacks/php/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Retry downloads during bootstrapping ([#226](https://github.com/heroku/buildpacks-php/pull/226))
 - Consistent handling of "non-URL" URLs (SSH-style or path) in `composer.json`/`composer.lock` repositories and package `source`s/`dist`s ([#105](https://github.com/heroku/buildpacks-php/issue/105), [#187](https://github.com/heroku/buildpacks-php/issue/187), [#208](https://github.com/heroku/buildpacks-php/pull/208))
+- Support object-style notation of repositories in composer.json ([#209](https://github.com/heroku/buildpacks-php/pull/209))
 
 ## [1.0.6] - 2025-07-31